project/requirements.txt
fastapi==0.115.12
pydantic-settings==2.8.1
uvicorn==0.34.1
asyncpg==0.30.0
tortoise-orm==0.25.0
aerich[toml]==0.8.2
httpx==0.28.1
pytest==8.3.5
gunicorn==22.0.0
pytest-cov==6.1.1
flake8==7.2.0
black==25.1.0
isort==6.0.1
python-dateutil
